v2: attempt to add proper stepper support
Damn, damn, I almost have it!

I can hard-code the speed and it works, but the RPM coming from RepG appears to be goofy, and I’m having trouble debugging it.

Here’s my diffs so far (I know it’s not working, but it’s something):

Hopefully tomorrow I can get it to work. I suspect that the:

from_host.read32(2)

On line 130 of Host.cc is where it’s broken. I think maybe read32() is broken, since I don’t see it actually used anywhere in the code except in a test case (!). I think it has endian issues.

To hook it up:
D10 is enable.
Direction is Pin 7 of that quadrature.
Step is Pin  8 of the quatrature.

In this photo: http://reprap.org/wiki/File:Cache-3456394835_0de9141966.jpg

Pin 7 is the bottom row, next to the right, and Pin 8 is in the top row directly above it.
Bottom right is ground, and top right is 5V.

You should connect ground to your stepper driver somewhere.

Hopefully I’ll get time to fix this tomorrow. If not, maybe someone else will. :)